# Make FNV hash compatible across endianness