@JanKrivanek 2023-12-22
> Since the compatible version seems to be faster anyway, this change seems reasonable, but what does that have to do with endianness? It seems like the span version and the not-span version do the same thing...

Casting byte ptr to char ptr leads to proper reordering of bytes (if needed) in memory by runtime (so that they are Little Endian, regardless of the actual storage arch)
When using the span version - we are basically handling the bytes on our own - so getting them in order as they are actually stored.

https://github.com/dotnet/msbuild/pull/9387#discussion_r1410414249 has more details