@KirillOsenkov 2024-01-09
We also have an Fnv1a implementation in the binlog writer:
https://github.com/dotnet/msbuild/blob/10f2f328089ccbc9f0c60cba527fbfe90d8294ff/src/Build/Logging/BinaryLogger/BuildEventArgsWriter.cs#L1314-L1342

See my analysis here:
https://github.com/KirillOsenkov/MSBuildStructuredLog/wiki/String-Hashing

Would be interesting to add this hash implementation to my benchmarks and compare:
https://github.com/KirillOsenkov/Benchmarks/blob/main/src/Tests/StringHash.Fnv.cs
https://github.com/KirillOsenkov/Benchmarks/blob/f2c45821c2cf7243b040d2c1db5904bab8134cf8/src/Tests/StringHash.cs#L73