@JanKrivanek 2024-01-09
> We also have an Fnv1a implementation in the binlog writer:
> 
> https://github.com/dotnet/msbuild/blob/10f2f328089ccbc9f0c60cba527fbfe90d8294ff/src/Build/Logging/BinaryLogger/BuildEventArgsWriter.cs#L1314-L1342
> 
> See my analysis here: https://github.com/KirillOsenkov/MSBuildStructuredLog/wiki/String-Hashing
> 
> Would be interesting to add this hash implementation to my benchmarks and compare: https://github.com/KirillOsenkov/Benchmarks/blob/main/src/Tests/StringHash.Fnv.cs https://github.com/KirillOsenkov/Benchmarks/blob/f2c45821c2cf7243b040d2c1db5904bab8134cf8/src/Tests/StringHash.cs#L73

Yeah - I 'borrowed' that one (`ComputeHash64Fast`)  :-) and quoted the source (your analysis):

https://github.com/dotnet/msbuild/pull/9489/files#diff-ed2e36ca70a3a73e4379cf4470aa8f4f492b961eabfaa15127b21efcff6706f1R58

Other than the `Span` usage the implementations are identical. However the `Span` usage didn't provide any significant benefit and required coditional compilation for .NET 3.5 - hence was pulled out.

Note though that currently this work is on waiting banch till the #9572 is fixed.