@rokonec 2024-01-12
[Blocking issue](https://devdiv.visualstudio.com/0bdbc590-a062-4c3f-b0f6-9383f67865ee/_workitems/edit/1937630) in AD.