@JanKrivanek 2024-02-08
Superseded by https://github.com/dotnet/msbuild/pull/9721